+++ src/ActionMonitor/Monitors/Monitor.php
@@ -3,6 +3,7 @@ namespace WPGatsby\ActionMonitor\Monitors;
 
 use GraphQLRelay\Relay;
 use WPGatsby\ActionMonitor\ActionMonitor;
+use WPGatsby\Admin\Preview;
 
 abstract class Monitor {
 
@@ -294,10 +295,42 @@ abstract class Monitor {
 		if ( $action_monitor_post_id !== 0 ) {
 
 			if ( isset( $args['preview_data'] ) ) {
+				$existing_preview_data = \get_post_meta(
+					$action_monitor_post_id,
+					'_gatsby_preview_data',
+					true
+				);
+
+				$manifest_id = Preview::get_preview_manifest_id_for_post(
+					get_post( $args['node_id'] )
+				);
+
+				$manifest_ids = [$manifest_id];
+
+				// if we have existing data, we want to merge our manifest id
+				// into any existing manifest ids
+				if ( $existing_preview_data && $existing_preview_data !== "" ) {
+					$existing_preview_data = json_decode( $existing_preview_data );
+
+					if ( $existing_preview_data->manifestIds ?? false ) {
+						$manifest_ids = array_unique(
+							array_merge(
+								$existing_preview_data->manifestIds,
+								$manifest_ids
+							)
+						);
+					}
+				}
+
+				// add manifest ids
+				$preview_data = json_decode( $args['preview_data'] );
+				$preview_data->manifestIds = $manifest_ids;
+				$preview_data = json_encode( $preview_data );
+				
 				\update_post_meta(
 					$action_monitor_post_id,
 					'_gatsby_preview_data',
-					$args['preview_data']
+					$preview_data
 				);
 			}